requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
dj-rest-auth==2.1.11
Django==3.2.7
django-environ==0.7.0
django-filter==21.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
kombu==5.2.1
MarkupSafe==2.0.1
packaging==21.0
Pillow==2.9.0
prompt-toolkit==3.0.22
psycopg2-binary==2.9.1
pyparsing==2.4.7
pytz==2021.1
redis==4.0.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==3.0.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
